ajax.js: simplify ajax URL logic
Save the entire URL to refresh, instead of just the `sn` query param.
This prevents us from needing to implement custom code to extract the
SN from the query string.

Reminder that we need to save the URL for cases where we change the
SN as part of an AJAX container update (rather than just a regular
auto-refresh) so that the subsequent auto-refreshes aren't accessing
outdated links.

A couple notes about implementation:

1. We don't need the sanity check that the URL has an `sn` parameter
   in `initRefresh`, because we only call `initRefresh` when we've
   explicitly enabled auto-refresh.

2. The `ajaxLink` function now _always_ makes an AJAX-enabled request.
   Previously, this was not happening for all AJAX containers (e.g.
   in chess_play.js), but testing suggests all behavior is unchanged.

3. Duplicate SN query params are no longer submitted as part of the
   auto-refresh request in `updateRefreshRequest`.

If we ever want to parse query strings again, we should use the built-
in class `URLSearchParams` instead of writing our own function.