UniGen: use single href for sector details
Change the link for selecting a sector to edit into a form with the
sector ID as a submit input. This allows us to reuse the same SN for
all sectors, further reducing the total number of SNs needed in the
UniGen page.

In the non-admin engine code, we do try to use container data instead
of request parameters when possible, because it completely prevents
falsification of form data; however, in the admin code, we generally
can assume that abuse is a non-issue.