Fix bad browser interaction with hidden field
When joining an alliance with open recruitment (i.e. no password), we
were using a hidden password field with an empty string value. Because
we did not make the field `readonly`, some browsers were auto-filling
this field with the password they associated with the website, which
then prevented them from joining the alliance ("Incorrect password!").

We could add the `readonly` property to the password field, but it is
perhaps best to remove this field entirely to avoid any client-side
issues. Therefore, we will assume an empty string password when the
password field is not submitted.