Ignore code smell AssignmentInCondition
In the case of the alliance_roster.php template, we really do want to
define `$Disabled` in the else-if conditional, because it is the only
way to avoid superfluous database calls in `SmrAccount::isDisabled`.

We may want to consider refactoring `SmrAccount::isDisabled`, so that
the query result is cached. But for now we leave it as is since it is
functional.