run-tests.yml: break steps into multiple jobs
The various types of CI tests (phpunit, phpstan, phpcs) do not depend
on each other, and can be run in parallel using separate jobs. Another
reason to use multiple jobs is that it's the only way to get separate
checks to display on a PR, which is helpful for quickly identifying
which component is failing.

These jobs all depend on the "build" job, where the Docker image is
actually built. Since jobs do not share environments, we must upload
the image and then download it in dependent jobs. We use a "composite"
action to avoid repeating setup steps in each job.

References:

* Share built image between jobs:
  https://github.com/docker/build-push-action/blob/master/docs/advanced/share-image-jobs.md

* Creating a composite action:
  https://docs.github.com/en/actions/creating-actions/creating-a-composite-action