docker-compose.yml: fix mysql-test dependencies
Of the various services based on the `smr-test` anchor, only `phpunit`
depends on `mysql-test`. Therefore, we remove it from the anchor and
specify it only for `phpunit` so that we don't unnecessarily download
the mysql image in the PHPStan and PHP_CodeSniffer CI jobs.

As a result of this change, the other `smr-test` services inherited
the "depends_on" fields of the `smr-cli` anchor. This helped clarify
that `smr-test` services should really inherit from the lower-level
anchor `smr-builder`.

Removing the `smr-cli` inheritance also removed the `networks` config
that `phpunit` was using (so that it could reach the `mysql` service).
Therefore, we also need to add the backend network to the `phpunit`
service.