CI: Use docker cache instead of artifacts
Instead of uploading the docker image as an artifact of the "build"
job, we use the `gha` cache option of docker/build-push-action.

Note that the `gha` cache is experimental:
https://github.com/docker/build-push-action/blob/master/docs/advanced/cache.md#github-cache

This is better because:

* Much faster than upload/download.
* Allows us to use build caching in the `build` job for free.
* No artifact management to worry about.

More details about the workflow speed:

* Build job is significantly faster:
  - Cache reduces build time from 2.5 minutes to ~30 seconds.
  - Cache upload is significantly faster (~5 seconds instead of 1 min),
    seemingly because the base images are already cached(?), so we do
    not need to upload the entire ~200MB image.

* Dependent jobs are about the same (~30 seconds to load image both
  from docker cache and from artifacts).

* Total time is 2.75 min (w/ no cache invalidation), compared to the
  fully serial build in a single job (3.5 min).

We could have kept the composite action, but at this point it didn't
seem worth the extra complexity. Hopefully there are better options for
reducing workflow duplication in the future.

Note that we need the extra `build-push-action` in each of the test
jobs, because it's the only way to get the docker layers out of the
cache (`cache-from: type=gha`) and into docker images (`load: true`).