Remove FFA game mode logic
This game mode was not fully implemented. The only restriction it made
was to exclude top racial ships (which is already implemented in the
"Hunter Wars" game mode). One might expect an FFA game mode to have a
rule about single-player alliances, but no such rule was in place.

We can re-implement the FFA game mode in the future if we want, but
for now the simplest solution is to remove it (as there are no plans
to use such a game mode).

Note that the FFA game mode is still selectable when creating a game,
but it will now be identical to the default game mode. This probably
should be prohibited, but we need to do so in a way that doesn't erase
the existence of the FFA game mode (since there are past games that
used this mode).