SmrPort: fix inconsistent state bug during PR
If a port lost a level on the same attack where it is destroyed, then
it would throw an error when trying to update the good amounts in the
database.

This was because we set all good amounts to 0 in `killPortByPlayer`,
populating the `goodAmountsChanged` property array with all good IDs;
then `removePortGood` would unset the lost good ID from `goodAmounts`,
but would not also remove it from `goodAmountsChanged`. This resulted
in the following error:

> TypeError: AbstractSmrPort::getGoodAmount(): Return value must be of type int, null returned

We now unset the lost good ID from additional property arrays to make
sure the class instance state is self-consistent.