Move USING_AJAX into the Session class
Defining `USING_AJAX` as a global constant made it hard to test, since
it cannot be modified once set. Moving it into the Session class as
the `ajax` property makes it easier to reset, and also much more clear
when it will be defined. In general, we should avoid global constants
when they are not configuration parameters.