Menu: reduce SemiWars headquarters duplication
In the Semi Wars game mode, we want the headquarters menu exactly the
same as a normal game, except without the Claim/Place Bounty pages.

To reduce code duplication, we can call into the parent method, but
indicate that we don't want the bouny pages with a new `addBountyPages`
argument (defaults to true).