Remove SmrPlayer class for Draft games
Move the logic for Draft games into the base SmrPlayer class. This
only impacts the `getHome` method.

Note that this introduces a dependence on SmrGame::getGame in
SmrPlayer::createPlayer, so for our tests, we need to use a real
SmrGame (and since we access it statically, we can't use a mock).