Add gameID attribute to SmrLocation class
This will be necessary to properly handle any SmrLocation logic that
depends on the current game type (e.g. restricted ship list for the
Hunter Wars game type).

We use `gameID = 0` when we don't care what the game type is.

Note that this also opens up the possibility of making the `location`
table customized for each game ID (to allow admins to make any changes
they want for a specific game, without changing the default config).