Disallow lib game type subdirectories
Class overrides based on game type have now all been removed, so we
now remove this concept from our custom autoloader.

The override design had the problem that it did not always load the
correct version of a class, and we couldn't change that version once
it had been loaded (due to how PHP class loading works).

It was also bad design to have multiple classes with the same name in
the same (global) namespace.

All the logic for non-default game types has been moved into the base
classes. While this does make it harder to see exactly how a game type
differs from the default, it is much safer and more explicit.

Maybe in the future we can use proper OOP and inheritance to implement
game type logic in a more encapsulated way. For example, if we were to
have `Player::getPlayer` create a `Draft\Player` or `SemiWars\Player`,
based on the game ID that was input, there would be no autoload race
conditions since the class names would not be duplicated.