Add gameID argument to bbifyMessage
We were setting the game ID for bbifyMessage in a dangerously implicit
way, by basing it on expected var fields (`game_id` or `GameID`) or the
session game ID as a fallback.

We may not always use exactly those var fields to define the game ID
associated with a page container, and it's not manageable to enumerate
all the possible variants due to the lack of encapsulation. Further,
those fields may not always be exactly the game ID we want to use for
bbifyMessage.

By passing the game ID directly to bbifyMessage, we ensure that we are
always using the intended game ID. We still use the current session
game ID as a fallback to reduce the size of the diff, but we may make
it a required option in the future for robustness.