current_sector.php: simplify logic of Refresh All message
We can use a boolean in the `current_sector.php` container to toggle
the display of the refresh countdown message after clicking the
"Refresh All" button.

Previously, we used a `[Force Check]` string in the general `message`
container field, which then was used to populate the `sector_message`
table with the same string. This was both fragile and overcomplicated.