ajax.js: refactor to reuse ajaxLink
Reuse the `ajaxLink` function in `updateRefreshRequest` instead of
having a separate low-level `$.get` call. This ensures that ajax calls
are handled in a consistent way.

Move the "refresh" logic out of `updateRefresh` and rename it to
`updateDisplay`; this function is now responsible only for modifying
the displayed contents of the page. The refresh logic is moved inside
the `ajaxLink` function.

Switch to named arguments in `ajaxLink` interface, to make it easier
to use the default values of some arguments.

No functionality changes.