Move ajax sanity check into Session class
We call `do_voodoo` possibly multiple times during a request, but we
really only need to perform the ajax sanity check once, when we know
what the request looks like.

Therefore, we can move the check into `Session::fetchVarInfo` and we
no longer need to store `$var['AJAX']` in a global constant (since
we won't need to remember it now for any page forwarding we do).

We could also demand that `$var['AJAX']` containers are _only_ called
with ajax, but that seems like an unnecessary constraint. All pages
where ajax is allowed would also work with a non-ajax call.

The call to `Template::display` now only depends on `$session->ajax`,
because whether `AJAX_CONTAINER` requires ajax or just allows it is
irrelevant. We want XML content if and only if request has `ajax=1`.