Session: fix back button error on page refresh
Prior to 5c5d1d914481, accessing an SN after it was already processed
would result in displaying the final Page of the processing chain,
since the initial processing would modify the stored Page associated
with that SN.

To fix a host of other issues, we no longer modify the `links` property
for a given SN during processing. This means that if we accessed an SN
again, it would start from scratch. If the SN corresponded to a non-
reusable Page, then it would throw a back button error.

In general, this is reasonable behavior, since we can't expect to store
all processed Pages in the session for all time (especially since Pages
will often store unique data that would make the session grow unbounded
in size).

However, one deficiency is that it prevented using F5 on a displayed
Page that resulted from processing a Processor link. For example, if
you click a link to move to another sector, that link will have an SN
associated with SectorMoveProcessor, a non-reusable page. As a result
of processing, you will be presented with a CurrentSector display page,
but the SN will still be associated with SectorMoveProcessor. So if you
then hit F5, it will try to process the SectorMoveProcessor link again,
which no longer exists.

We could redirect the user to a new URL with an SN that corresponds to
its resulting display page, which would prevent almost all back button
errors, but that would be tantamount to storing a potentially unbounded
number of SNs in the session. This is perhaps something to keep in mind
if we ever want to fully redesign the Session.

As a compromise, we can fix the more practical issue with the F5 back
button error only. We do this by detecting the following scenario:

1. The requested SN is the same as the one we just processed (i.e. URL
   has not changed).

2. The SN corresponds to a Page link that no longer exists (i.e. it was
   a non-reusable Page, likely a Processor page).

3. It is not an ajax request.

In this case, we use the `$lastPage`, which _is_ the display Page that
corresponds to the end of the processing chain for the given SN, and
use it as the `currentPage`. That is, we're displaying the last Page
we displayed, without doing all the processing that was done the first
time it was loaded (since those processing Pages are likely to be
non-reusable). This will allow users to press F5 without getting any
back button errors.

Note that we also restrict newly generated SNs from being the same as
the SN we just processed. Since we free up the SN immediately for non-
reusable pages, in the unlikely case that a different Page picked the
same SN, reloading the Page (both F5 or ajax) would result in the Page
content changing!