SmrPlayer: avoid duplication in HOF calls
In `killPlayerByPlayer`, we had a lot of duplication for killing NPCs
vs. other players. The only difference is that for killing NPCs we
insert an "NPC" tag into the HOF category.

One category was changed for consistency:

OLD => Killing:Experience:Lost By NPCs Killed
NEW => Killing:NPC:Experience:Lost By Traders Killed

We also add an "NPC" tag for dying to NPCs, now that it is possible!
It uses the same deduplication technique as for killing NPCs.