Add StoredDestination data class
This class stores the data from the `player_stored_sector` table, which
was previously only stored in an array in the SmrPlayer class. Using a
data class provides much more robust storage and access, as it codifies
the interface to the object.

I debated making some of the properties _not_ readonly, so that an
existing object could be modified, e.g. changing the position offsets
in `SmrPlayer::moveDestinationButton`. This might simplify some things,
but it seemed more robust and not too inconvenient to replace the
object with an entirely new one. As a result, we can ensure that no
consumer is able to modify the StoredDestination objects stored in the
SmrPlayer.

The original intention of this change was to simply allow for storing a
sector without a label, and to display it properly (i.e. no dash). We
implement this logic in the `StoredDestination::getDisplayName` method.