CreateGalaxies: don't hard-code racial galaxy names
We can make the default galaxy name list more robust by querying the
list of playable races, and using those as the first N galaxy names.

Note that we remove the bogus "0" at the start of the list of names,
and then decrement the index when accessing it (since the `game_galaxy`
table uses a 1-based index).