Fix alliance invitations and closed recruitment
When an alliance has closed recruitment, it is supposed to be possible
to accept a direct invitation. We add an option to `getJoinRestriction`
to allow for overriding the recruitment status check, so that accepting
an alliance invitation will work with closed recruitment again.

It's not clear if this was ever working as intended before now.