AllianceMessageBoard: fix Author display
The displayed Author of a thread was not the creator of the thread, as
one would expect. Instead, we were using the player with the smallest
account ID who had participated in the thread(!). The behavior had been
wrong for the entire git history.

We now use a window function in the query to select the account ID tied
to the author of the first post in the thread to determine authorship.

Technical SQL note
==================

We couldn't just change

  min(sender_id)

to

  FIRST_VALUE(sender_id) OVER (PARTITION BY thread_id ORDER BY reply_id)

and leave the rest of the query unchanged. Even though doing so looked
like it worked properly in the exact case we needed, it was only
coincidentally correct. The window function would not work as expected
when combined with the GROUP BY clause due to precedence order, and the
fact that MySQL returns an ill-defined value if a returned column in a
GROUP BY is multi-valued. This is why we needed separate CTE's to run
the window function first.

See https://stackoverflow.com/a/52412604/8834658.

A simpler solution would be to add an `author_id` column to the
`alliance_thread_topic` table, rather than using the author of the
first post in the topic. Such a solution would also fix the issue of
the author of the topic changing when the first post is deleted.