login.php: remove updateLastLogin call
Require that the user gets to the LoginProcessor page before updating
the login time. This is the last processor page before we forward to
the GamePlay display page, and deferring the login time update ensures
that the user saw announcements and/or changelogs. If they manage to
skip those pages (e.g. clicking on left panel links, which me may want
to consider hiding for these pages), this ensures that the next time
they log in, they'll still be directed to the content they missed.