Swap Buy/Sell locations in port display
The vertical ordering of Buy and Sell goods on the Current Sector and
Local/Galaxy Map is now swapped to match the ordering of goods upon
entering the port.

This ordering was selected (rather than the other way around), because
it's more friendly to new players to see the goods you can buy first,
as that is the first thing they will do.

Also synchronized the CSS and HTML of the two locations this display
is produced. (I would have made it a sub-template, but I worried that
it might be too expensive in the Galaxy Map.) The Buy/Sell rows no
longer have vertical space between them in the Current Sector.