CurrentSector: fix empty TurnsMessage display
Since `TurnsMessage` is always a string, it was always passing the
`isset` check to display it. The display included an extra line break,
which was then causing extra vertical whitespace when there was no
TurnsMessage to display.

To fix this, we only set the `TurnsMessage` template var if the message
is non-empty.