Fix regression in Player::removeUnderAttack
The behavior introduced during the Page refactor in 1361b05b3d37 did
not work as intended. In cases where the player was attacked between
ajax updates, the under attack warning would not be displayed, because
the `underAttack` method of the Page had already been set to false by
the original page load.

The new behavior reverts to essentially the previous logic, except that
we move it from a SmrPlayer method to a Page method, since this allows
us to make the `underAttack` Page property protected.