Fix skipped Discord/CommandTest unit test case
The reason this test case was skipped was because it called a global
function `logException`, which could not be mocked (by PHPUnit design).
As a result, it would echo the log message to the test output.

The easiest fix is to wrap the call in a `Command` method, which could
be mocked. It doesn't feel great to modify an interface solely for the
purposes of testing, but it is less offensive than not testing at all.

In the future, using Monolog and an injected logger instance may be a
more robust solution.