npc.php: fix issue causing no-op trading
Some logic added to ShopGoodsProcessor in e280e9ad93 resulted in NPCs
performing a no-op when executing the container, because they were
being forwarded to the ShopGoodsNegotiate display page due to having
`bargainNumber` set to 0 (the default). By setting a non-zero value for
this argument, we simulate actually processing a transaction rather
than requesting to negotiate. This confusion is partly because the
ShopGoodsProcessor page is overloaded and needs to be broken up into
separate processor pages.