HallOfFame: fix partial matching of categories
If the requested category matches the first part of another category,
it would be displayed in the wrong place.

For example, in the "Killing -> NPC" category, the following sub-
categories were displayed:

* Experience
* Money
* Alignment
* NPC Kills

However, the full category types for these are, respectively:

* Killing:NPC:Experience
* Killing:NPC:Money
* Killing:NPC:Alignment
* Killing:NPC Kills

Notice how "NPC Kills" is not in the "Killing:NPC" category, instead
it is a subcategory of "Killing" only.

By adding a ":" to the end of the type to filter against, we ensure
that no partial matches like this are displayed.