DatabaseIntegrationTest: fix getDbBytes failure
This test always works reproducibly when the test database is first
created (and therefore it never fails in the GHA workflow), but it
starts reporting discrepancies locally 24 hours after the test database
is initialized (since we usually skip reinitializing the database when
running tests locally to speed up the process). E.g.:

> Failed asserting that 730800 is identical to 735034.

The failure does not correspond to the contents of the database being
different or wrong (the checksums are all correct). The culprit turned
out to be `ship_*` tables in SmrShipIntegrationTest that we insert into
and then only DELETE. I thought that because we deleted all the rows we
added, it was safe to not include these tables in `tablesToTruncate`.
We must use TRUNCATE to reclaim the storage for these tables (if we're
not going to reinitialize the database for each test execution).

The reason this issue only manifested after 24 hours is explained here:
https://dev.mysql.com/doc/refman/8.0/en/information-schema-tables-table.html

> Columns in TABLES that represent table statistics hold cached values.
> The `information_schema_stats_expiry` system variable defines the
> period of time before cached table statistics expire. The default is
> 86400 seconds (24 hours).

So the storage for these tables was actually growing after the first
time the SmrShipIntegrationTest was run, but the cached statistics
(from the initial CHECKSUM analysis) were not updated until 24h later.

Adding these tables to `tablesToTruncate` is also important to avoid
leaving the database in an unsafe state if the DELETE operations are
skipped because an earlier part of the test fails. We have a failsafe
for this in `BaseIntegrationSpec::checkTables`, but we shouldn't be
relying on that.