Replace trade good array with TradeGood class
This fixes PHPStan warnings due to the imprecise typehint that we were
using: `array<string, int|bool>`. Alternatively, we could have used an
explicit array shape typehint, but the immutability of the data class
adds a lot of robustness.

Since TradeGood is a data class, all properties are readonly and public
so that we don't need to implement any getters.

Note that there is no TradeGood instance that represents GOODS_NOTHING,
i.e. it's only used for real trade goods. Many places that were using
Globals::getGoodName should not have been allowed to use GOODS_NOTHING,
so by using the TradeGood::name property, we make that restriction
explicit.

The only place where GOODS_NOTHING is actually being used is in the
Routes classes, where GOODS_NOTHING is used to scaffold one-way routes.
Therefore, we add the method OneWayRoute::getGoodName, which allows
for a "Nothing" good name.

Rather than just providing the TradeGood image source, we add a method
`getImageHTML` that provides the full HTML `<img>` element. This is
useful because it reduces a lot of template duplication. We had to
tweak some of the css styles to get everything to display properly.