UniGen: use HTML5 date type for join/start/end times
HTML5 date types are submitted to the server as "Y-m-d", so they can be
passed directly to the `DateTime` constructor. If the wrong format is
given, it will simply throw an Exception (unlike with the previous use
of `DateTime::createFromFormat`, which returns false on invalid input).

Since this is for an Admin Tool, we do not expect users to be bypassing
the HTML5 validation (and if it does it's just an Exception), so we do
not need to bother with UserErrors on invalid input.

The displayed date format will be locale/browser-dependent, so no use
in giving a separate format hint. Most browsers will provide a nice
calendar button to help out.

Fixes PHPStan warning in EditGameProcessor:

> Cannot call method getTimestamp() on DateTime|false.