SmrPlayer: add hasCustomShipName type narrowing doc
Use `@phpstan-assert-if-true` type narrowing to teach PHPStan that,
in SmrPlayer, if `hasCustomShipName` is true, then `getCustomShipName`
will not return false (i.e. it will return a string). This is a useful
strategy for more contextual interpretation of union return types,
especially for hasser/getter method pairs.

See https://phpstan.org/writing-php-code/narrowing-types#custom-type-checking-functions-and-methods.