IpViewResults: use explode instead of preg_split
Since all we want to do is split a string on a delimiter, and then trim
the results, we can do that much more explicitly with the `explode` and
`trim` methods. With `preg_split`, we would have needed to validate
that the result did not return false (which also would have prevented
us from using array destructuring as well).

Fixes the following PHPStan warnings:

>  158    Cannot use array destructuring on array<int, string>|false.
>  187    Parameter #1 $array of method Smr\Database::escapeArray() expects array, array<int, string>|false given.
>  195    Parameter #1 $array of method Smr\Database::escapeArray() expects array, array<int, string>|false given.

Unrelatedly, fixed an issue with player names not being escaped
properly for display.