Plotter: fix validation of getX input
We cannot pass an int value of `$X` to the `str2int` method, because it
only accepts strings. This regression was introduced in 7330d7aa1850,
and was causing an error when accepting missions (since that's the only
place we currently pass an int `$X`).

We may want to consider changing `str2int` to allow other types, e.g.
int or float or even mixed, as the implementation works for any type
(but then obviously we'd need to rename the function).

Fixes the PHPStan warning:

> Parameter #1 $val of function str2int expects string, int|string given.