Replace array with ShipIllusion class
This data class is used in the SmrShip class as a safer alternative to
explicitly documenting the shape of the prior `illusionShip` array.

We replace the following methods with public readonly property access
on the ShipIllusion class:

* getIllusionShipID() -> shipTypeID
* getIllusionAttack() -> attackRating
* getIllusionDefense() -> defenseRating
* getIllusionShipName() -> getName()

Also rename `getIllusionShip` to `getIllusion` for brevity.

Use `@phpstan-assert-if-true` to codify the relationship between the
`hasActiveIllusion` and `getIllusion` SmrShip methods, so that PHPStan
knows when we'll get a ShipIllusion object rather than false.

Fixes PHPStan warnings:

> Cannot access offset 'ID' on array<string, int>|false.
> Cannot access offset 'ID' on array<string, int>|false.
> Cannot access offset 'Attack' on array<string, int>|false.
> Cannot access offset 'Defense' on array<string, int>|false.