Replace hardware type array with HardwareType class
The Globals::getHardwareTypes returned an array whose shape was not
explicitly typehinted, which resulted in a lot of PHPStan warnings due
to some fields being strings and others being ints.

The HardwareType class is a data class with public readonly properties
for the hardware type ID, name, and cost. This is a better alternative
to simply typehinting the explicit shape of the previous data array.