Add PlayerLevel class to replace Globals method
The Globals::getLevelRequirements method is replaced by this class,
which stores the level ID, name, and required exp as public readonly
properties.

Replace some of the SmrPlayer methods with PlayerLevel methods that
can be chained with the SmrPlayer::getLevel method, e.g.

  $player->getLevel()->next()->expRequired

instead of

  $player->getNextLevelExperience()

This simplifies the SmrPlayer interface, and fixes PHPStan warnings
that resulted from not explicitly documenting the array shape of the
prior level array.

> Method AbstractSmrPlayer::getNextLevelExperience() should return int but returns int|string.
> Method AbstractSmrPlayer::getThisLevelExperience() should return int but returns int|string.
> Property AbstractSmrPlayer::$level (int|null) does not accept int|false.
> Method AbstractSmrPlayer::getLevelID() should return int but returns int|false.
> Method AbstractSmrPlayer::getLevelName() should return string but returns int|string.
> Method AbstractSmrPlayer::getMaxLevel() should return int but returns int|false.