Globals: remove static Database storage
There is no performance gain with storing the Database as a static
property of the Globals class.

This fixes an issue where the Database is being mocked in a unit test,
and then that mock object is stored in Globals with no way to change
it for other tests. Using `Database::getInstance` each time it's needed
allows us to mock locally to a test.