Add WeaponDamageData docstring type alias
Narrow the typehint for "weapon data" arrays by using an explicit array
shape, since each field can only ever be a single type.

This fixes the following PHPStan warnings:

> Parameter #1 $damage of method AbstractSmrPort::takeDamageToShields() expects int, bool|int given.
> Parameter #1 $damage of method AbstractSmrShip::takeDamageToShields() expects int, bool|int given.
> Parameter #1 $damage of method SmrForce::takeDamageToMines() expects int, bool|int given.
> Parameter #1 $damage of method SmrPlanet::takeDamageToShields() expects int, bool|int given.

Ideally, we would convert this array to a data class, but the combat
templates are sufficiently fragile that I wanted to make the smallest
changes possible to fix the warnings, which meant only changing the
docstrings.

To avoid repeating a complex array shape in many places, we use the
PHPStan `typeAliases` configuration option, which allows us to specify
the type once, and then use it in docstrings as if it were a real type.