Add Bounty data class
Move the handling of `bounty` database records out of SmrPlayer and
into a dedicated Bounty class. The SmrPlayer class will still hold an
array of Bounty objects, and try to update them when SmrPlayer::update
is called.

As a result, the SmrPlayer interface has been substantially simplified.
Many method have been removed:

* getBountiesData
* getBounty
* hasBounty
* setBounty
* getCurrentBountyAmount
* setCurrentBountyAmount
* increaseCurrentBountyAmount
* decreaseCurrentBountyAmount
* getCurrentBountySmrCredits
* setCurrentBountySmrCredits
* increaseCurrentBountySmrCredits
* decreaseCurrentBountySmrCredits

The SmrPlayer class now only has methods to return a bounty, which can
then be manipulated by calling methods on the Bounty class to, e.g.,
increase the credits or SMR credits.

Other SmrPlayer methods now return Bounty types (or `array<Bounty>`)
instead of arrays (or `array<array<mixed>>`):

* getClaimableBounties
* createBounty
* getActiveBounty (renamed from getCurrentBounty)

The name change in methods from `*CurrentBounty` to `*ActiveBounty` is
because "active" makes it more clear that it's an outstanding bounty.