Add method SmrGame::getLastSectorID
Refactor duplicate logic for getting the last sector ID in the game
into a single method. Validate that there is at least one sector.

There is not an obvious place to put this function, but SmrGame seems
reasonable. This may not be the most efficient implementation either.
If not all SmrGalaxy objects are cached already, it may be faster to
perform a single query of the `sector` table instead.

Fixes PHPStan warning:

> Cannot call method getEndSector() on SmrGalaxy|false.