Do not allow false return from getCachedPort methods
The `getCachedPort` methods of SmrPort and SmrSector no longer return
false. Instead, they will throw a CachedPortNotFound exception if the
port does not exist in the database.

The reason for this is so that we don't have to check that the return
value is an SmrPort when we have already checked with `hasCachedPort`.

(This could have been an ideal case for `@phpstan-assert-if-true`, but
this doesn't work for hassers/getters when they take an argument, i.e.
SmrPlayer in this case.)

Fixes PHPStan warnings of the form:

> Cannot call method getLevel() on SmrPort|false.
> Cannot call method addCachePorts() on SmrPort|false.