Database: validate query result for read method
We added validation of the result for the `write` method during the
major Database refactor in 7054e3d9c18f, but the `read` method was not
similarly modified.

For the `read` method, we expect a `mysqli_result`. If it returns bool
instead, then it failed (false) or was the wrong query type (true).

Fixes PHPStan warning:

> Parameter #1 $dbResult of class Smr\DatabaseResult constructor expects mysqli_result, bool|mysqli_result given.